mail controller complete

+++ back-end/src/Controller/MailController.php
@@ -3,7 +3,6 @@
 namespace App\Controller;
 
 use App\Entity\Mail;
-use App\Entity\User;
 use App\Repository\MailRepository;
 use App\Repository\UserRepository;
 use DateTime;
@@ -18,10 +17,14 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class MailController extends AbstractController
 {
     #[Route('{microsoftId}', methods: ['GET'])]
-    public function index(MailRepository $mailRepository): JsonResponse
+    public function index(MailRepository $mailRepository, UserRepository $userRepository, string $microsoftId): JsonResponse
     {
-        //TODO
-        $mails = $mailRepository->findAll();
+
+        $user = $userRepository->findOneBy(['microsoftId' => $microsoftId]);
+        if (!$user) {
+            return $this->json(['error' => 'User not found'], Response::HTTP_NOT_FOUND);
+        }
+        $mails = $mailRepository->findBy(['user' => $user]);
 
         $data = array_map(function (Mail $mail) {
             return [
@@ -30,24 +33,27 @@ class MailController extends AbstractController
                 'sender' => $mail->getSender(),
                 'recipients' => $mail->getRecipients(),
                 'body' => $mail->getBody(),
-                'receivedAt' => $mail->getReceivedAt()->format('Y-m-d H:i:s'),
-                'user' => $mail->getUser()?->getId(),
+                'receivedAt' => $mail->getReceivedAt()->format('Y-m-d H:i:s')
             ];
         }, $mails);
 
         return $this->json($data);
     }
 
-    #[Route('/{id}', methods: ['GET'])]
-    public function show(Mail $mail): JsonResponse
+    #[Route('/{messageId}', methods: ['GET'])]
+    public function show(Mail $mail, MailRepository $mailRepository, string $messageId): JsonResponse
     {
+        $mail = $mailRepository->findOneBy(['messageId' => $messageId]);
+        if (!$mail) {
+            return $this->json(['error' => 'Mail not found'], Response::HTTP_BAD_REQUEST);
+        }
         return $this->json([
             'id' => $mail->getId(),
             'subject' => $mail->getSubject(),
             'sender' => $mail->getSender(),
             'recipients' => $mail->getRecipients(),
             'body' => $mail->getBody(),
-            'receivedAt' => $mail->getReceivedAt()->format('Y-m-d H:i:s'),
+            'receivedAt' => $mail->getReceivedAt()?->format('Y-m-d H:i:s'),
             'user' => $mail->getUser()?->getId(),
         ]);
     }
@@ -59,17 +65,31 @@ class MailController extends AbstractController
         UserRepository $userRepo
     ): JsonResponse {
         $data = json_decode($request->getContent(), true);
-
-        $user = $userRepo->find($data['user_id'] ?? null);
+        if (!$data['user']['microsoftId']) {
+            return $this->json(['error' => 'missing [\'user\'][\'microsoftId\'] '], Response::HTTP_BAD_REQUEST);
+        }
+        if (!$data['sender']) {
+            return $this->json(['error' => 'missing [\'sender\'] '], Response::HTTP_BAD_REQUEST);
+        }
+        if (!$data['recipients']) {
+            return $this->json(['error' => 'missing [\'recipients\'] '], Response::HTTP_BAD_REQUEST);
+        }
+        if (!$data['receivedAt']) {
+            return $this->json(['error' => 'missing [\'receivedAt\'] '], Response::HTTP_BAD_REQUEST);
+        }
+        if (!$data['messageId']) {
+            return $this->json(['error' => 'missing [\'messageId\'] '], Response::HTTP_BAD_REQUEST);
+        }
+        $user = $userRepo->findOneBy(['microsoftId' => $data['user']['microsoftId']]);
         if (!$user) {
-            return $this->json(['error' => 'User not found'], Response::HTTP_BAD_REQUEST);
+            return $this->json(['error' => 'User not found'], Response::HTTP_NOT_FOUND);
         }
 
         $mail = new Mail();
-        $mail->setSubject($data['subject'] ?? '');
+        $mail->setSubject($data['subject'] ?? 'Aucun objet');
         $mail->setSender($data['sender'] ?? '');
         $mail->setRecipients($data['recipients'] ?? []);
-        $mail->setBody($data['body'] ?? '');
+        $mail->setBody($data['body'] ?? 'Aucun contenu');
         $mail->setReceivedAt(new DateTime($data['receivedAt'] ?? 'now'));
         $mail->setMessageId($data['messageId'] ?? uniqid());
         $mail->setUser($user);
@@ -80,12 +100,59 @@ class MailController extends AbstractController
         return $this->json(['id' => $mail->getId()], Response::HTTP_CREATED);
     }
 
-    #[Route('/{id}', methods: ['DELETE'])]
-    public function delete(Mail $mail, EntityManagerInterface $em): JsonResponse
+    #[Route('/{messageId}', methods: ['DELETE'])]
+    public function delete(Mail $mail, EntityManagerInterface $em, MailRepository $mailRepository, string $messageId): JsonResponse
     {
+        $mail = $mailRepository->findOneBy(['id' => $messageId]);
+        if (!$mail) {
+            return $this->json(['error' => 'Mail not found'], Response::HTTP_NOT_FOUND);
+        }
         $em->remove($mail);
         $em->flush();
 
         return $this->json(null, Response::HTTP_NO_CONTENT);
+    }
+
+    #[Route('/message/{messageId}', name: 'update_by_message_id', methods: ['PUT', 'PATCH'])]
+    public function updateMailByMessageId(
+        string $messageId,
+        Request $request,
+        MailRepository $mailRepository,
+        EntityManagerInterface $em
+    ): JsonResponse {
+        $mail = $mailRepository->findOneBy(['messageId' => $messageId]);
+
+        if (!$mail) {
+            return $this->json(['error' => 'Mail not found'], Response::HTTP_NOT_FOUND);
+        }
+
+        $data = json_decode($request->getContent(), true);
+
+        if (isset($data['subject'])) {
+            $mail->setSubject($data['subject']);
+        }
+
+        if (isset($data['body'])) {
+            $mail->setBody($data['body']);
+        }
+
+        if (isset($data['recipients']) && is_array($data['recipients'])) {
+            $mail->setRecipients($data['recipients']);
+        }
+
+        if (isset($data['receivedAt'])) {
+            try {
+                $mail->setReceivedAt(new DateTime($data['receivedAt']));
+            } catch (\Exception $e) {
+                return $this->json(['error' => 'Invalid receivedAt format'], Response::HTTP_BAD_REQUEST);
+            }
+        }
+
+        $em->flush();
+
+        return $this->json([
+            'message' => 'Mail updated',
+            'messageId' => $mail->getMessageId(),
+        ]);
     }
 }
